requirements.txt
# Dependencias principales
# Navegación web automatizada
playwright==1.40.0
selenium==4.15.2

# Análisis de HTML y scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Procesamiento de lenguaje natural
nltk==3.8.1
scikit-learn==1.3.2
spacy==3.7.2

# Backend web
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0

# Procesamiento de datos
pandas==2.1.1
numpy==1.26.1

# Visualización
matplotlib==3.8.1
seaborn==0.13.0

# Utilidades
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
pillow==10.1.0

# Exportación de datos
openpyxl==3.1.2
